+++ qiskit/providers/ibmq/accountprovider.py
@@ -15,6 +15,7 @@
 """Provider for a single IBM Quantum Experience account."""
 
 import logging
+from typing import Dict, List, Callable, Optional, Any
 from collections import OrderedDict
 
 from qiskit.providers import BaseProvider
@@ -26,6 +27,9 @@ from qiskit.validation.exceptions import ModelValidationError
 from .api_v2.clients import AccountClient
 from .circuits import CircuitsManager
 from .ibmqbackend import IBMQBackend, IBMQSimulator
+from .credentials import Credentials
+from .providerbackends import ProviderBackends
+
 
 logger = logging.getLogger(__name__)
 
@@ -33,8 +37,13 @@ logger = logging.getLogger(__name__)
 class AccountProvider(BaseProvider):
     """Provider for a single IBM Quantum Experience account."""
 
-    def __init__(self, credentials, access_token):
+    def __init__(self, credentials: Credentials, access_token: str) -> None:
         """Return a new AccountProvider.
+
+        The ``provider_backends`` attribute can be used to autocomplete
+        backend names, by pressing ``tab`` after
+        ``AccountProvider.provider_backends.``. Note that this feature may
+        not be available if an error occurs during backend discovery.
 
         Args:
             credentials (Credentials): IBM Q Experience credentials.
@@ -55,7 +64,15 @@ class AccountProvider(BaseProvider):
         # access.
         self._backends = None
 
-    def backends(self, name=None, filters=None, **kwargs):
+        self.provider_backends = ProviderBackends(self)
+
+    def backends(
+            self,
+            name: Optional[str] = None,
+            filters: Optional[Callable[[List[IBMQBackend]], bool]] = None,
+            timeout: Optional[float] = None,
+            **kwargs: Any
+    ) -> List[IBMQBackend]:
         """Return all backends accessible via this provider, subject to optional filtering.
 
         Args:
@@ -63,6 +80,7 @@ class AccountProvider(BaseProvider):
             filters (callable): more complex filters, such as lambda functions
                 e.g. AccountProvider.backends(
                     filters=lambda b: b.configuration['n_qubits'] > 5)
+            timeout (float or None): number of seconds to wait for backend discovery.
             kwargs: simple filters specifying a true/false criteria in the
                 backend configuration or backend status or provider credentials
                 e.g. AccountProvider.backends(n_qubits=5, operational=True)
@@ -72,7 +90,7 @@ class AccountProvider(BaseProvider):
         """
         # pylint: disable=arguments-differ
         if self._backends is None:
-            self._backends = self._discover_remote_backends()
+            self._backends = self._discover_remote_backends(timeout=timeout)
 
         backends = self._backends.values()
 
@@ -86,15 +104,18 @@ class AccountProvider(BaseProvider):
 
         return filter_backends(backends, filters=filters, **kwargs)
 
-    def _discover_remote_backends(self):
+    def _discover_remote_backends(self, timeout: Optional[float] = None) -> Dict[str, IBMQBackend]:
         """Return the remote backends available.
+
+        Args:
+            timeout (float or None): number of seconds to wait for the discovery.
 
         Returns:
             dict[str:IBMQBackend]: a dict of the remote backend instances,
                 keyed by backend name.
         """
         ret = OrderedDict()
-        configs_list = self._api.available_backends()
+        configs_list = self._api.available_backends(timeout=timeout)
         for raw_config in configs_list:
             # Make sure the raw_config is of proper type
             if not isinstance(raw_config, dict):
@@ -124,7 +145,7 @@ class AccountProvider(BaseProvider):
         return ret
 
     @staticmethod
-    def _deprecated_backend_names():
+    def _deprecated_backend_names() -> Dict[str, str]:
         """Returns deprecated backend names."""
         return {
             'ibmqx_qasm_simulator': 'ibmq_qasm_simulator',
@@ -133,7 +154,7 @@ class AccountProvider(BaseProvider):
             }
 
     @staticmethod
-    def _aliased_backend_names():
+    def _aliased_backend_names() -> Dict[str, str]:
         """Returns aliased backend names."""
         return {
             'ibmq_5_yorktown': 'ibmqx2',
@@ -142,10 +163,10 @@ class AccountProvider(BaseProvider):
             'ibmq_20_austin': 'QS1_1'
             }
 
-    def __eq__(self, other):
+    def __eq__(self, other: 'AccountProvider') -> bool:
         return self.credentials == other.credentials
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         credentials_info = "hub='{}', group='{}', project='{}'".format(
             self.credentials.hub, self.credentials.group, self.credentials.project)